Extra collection range | Rebalance effects

+++ Assets/Scripts/Passive.cs
@@ -7,5 +7,5 @@ public class Passive : ScriptableObject
     public string passiveName;
     public string passiveDescription;
     public int extraBurrowDamage;
-    public int extraCollectionRange;
+    public float extraCollectionRange;
 }
